+++ src/ui/controls/slider/index.tsx
@@ -10,6 +10,7 @@ interface SliderProps {
     step: number;
     formatValue: (value: number) => string;
     onChange: (value: number | null) => void;
+    onPreview?: (value: number | null) => void;
 }
 
 interface SliderStore {
@@ -41,6 +42,7 @@ export default function Slider(props: SliderProps) {
 
     function onRootCreate(rootNode: HTMLElement) {
         rootNode.addEventListener('touchstart', onPointerDown, {passive: true});
+        rootNode.addEventListener('wheel', onWheel, {passive: false});
     }
 
     function saveTrackNode(el: HTMLElement) {
@@ -114,6 +116,9 @@ export default function Slider(props: SliderProps) {
             const value = getEventValue(e);
             store.activeValue = value;
             context.refresh();
+
+            const {onPreview} = store.activeProps;
+            onPreview?.(value);
         }
 
         function onPointerUp(e: MouseEvent | TouchEvent) {
@@ -198,7 +203,6 @@ export default function Slider(props: SliderProps) {
             class={{'slider': true, 'slider--active': store.isActive}}
             oncreate={onRootCreate}
             onmousedown={onPointerDown}
-            onwheel={onWheel}
         >
             <span
                 class="slider__track"