+++ src/ui/icons/phone-icon.tsx
@@ -1,0 +1,10 @@
+import {m} from 'malevic';
+
+export function PhoneIcon(): Malevic.Child {
+    return <svg viewBox="0 0 16 16">
+        <g stroke="currentcolor" stroke-width="1.5" fill="none">
+            <rect x="3" y="1" width="10" height="14" rx="2" ry="2" />
+            <line x1="6" y1="13" x2="10" y2="13" stroke-width="1" />
+        </g>
+    </svg>;
+}