+++ tasks/code-style.js
@@ -1,5 +1,5 @@
 // @ts-check
-import prettier from 'prettier';
+import {format} from 'prettier';
 import {getDestDir} from './paths.js';
 import {PLATFORM} from './platform.js';
 import {createTask} from './task.js';
@@ -22,7 +22,7 @@ const extensions = ['html', 'css', 'js'];
 async function processAPIBuild() {
     const filepath = 'darkreader.js';
     const code = await readFile(filepath);
-    const formatted = await prettier.format(code, {
+    const formatted = await format(code, {
         ...options,
         filepath,
     });
@@ -36,7 +36,7 @@ async function processExtensionPlatform(platform) {
     const files = await getPaths(extensions.map((ext) => `${dir}/**/*.${ext}`));
     for (const file of files) {
         const code = await readFile(file);
-        const formatted = await prettier.format(code, {
+        const formatted = await format(code, {
             ...options,
             filepath: file,
         });