+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+# 4.9.95 (October 2, 2024)
+
+- Fixed Google Calendar scroll bars.
+- Fixed automation panel.
+- Added links to mobile apps.
+
+# 4.9.94 (September 24, 2024)
+
+- API: Fixed TypeScript builds.
+- Updated experimental design for Microsoft Edge for Android.
+
+# 4.9.93 (September 20, 2024)
+
+- Experimental design for Microsoft Edge for Android.
+
+# 4.9.92 (September 10, 2024)
+
+- Detect website's dark theme by default.
+- Use Filter mode for particular websites by default (Google Docs, Microsoft Office).
+- Restored scrollbar colors.
+- Fixed breaking on default selection color.
+
 # 4.9.90 (September 1, 2024)
 
 - Fixed "Aw, Snap" in Google Search (Chrome and Edge on Windows).